requirements.txt
google-genai 
Pillow
fastapi[standard]
python-dotenv
openai
pytest
pytest-asyncio
pytest-cov
pytest-html
pytest-xdist
httpx